@@ -1,7 +1,7 @@
 name: sqflite_common_ffi_web
 repository: https://github.com/tekartik/sqflite/tree/master/packages_web/sqflite_common_ffi_web
 description: Sqflite web implementation using sqlite3 ffi and sqlite3 wasm.
-version: 0.4.0
+version: 0.4.3+1
 funding:
   - https://github.com/sponsors/alextekartik
 
@@ -14,28 +14,27 @@ topics:
   - database
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   path: '>=1.7.0 <3.0.0'
-  sqflite_common_ffi: '>=2.3.0-1 <4.0.0'
-  sqflite_common: '>=2.5.0-1 <4.0.0'
-  sqlite3: '^2.0.0'
+  sqflite_common_ffi: '>=2.3.3 <4.0.0'
+  sqflite_common: '>=2.5.4 <4.0.0'
+  sqlite3: '>=2.3.0 <4.0.0'
   http: '>=0.13.4 <2.0.0'
-  js: '>=0.6.4 <2.0.0'
+  web: '>=0.5.0 <2.0.0'
   synchronized: '>=3.0.0+3 <5.0.0'
   process_run: '>=0.12.3+2 <2.0.0'
-  dev_test: '>=0.15.7+2 <2.0.0'
+  dev_build: '>=0.16.3+1 <2.0.0'
   pub_semver: '>=2.1.2 <4.0.0'
-  io: '>=1.0.3 <3.0.0'
   args: '>=2.3.1 <4.0.0'
   meta: '>=1.8.0 <3.0.0'
 dev_dependencies:
   test:
-  build_runner: '>=2.1.4 <4.0.0'
-  build_web_compilers: '>=3.2.1 <5.0.0'
-  build_daemon: '>=3.0.0 <5.0.0'
-  lints: '>=2.0.0 <4.0.0'
+  build_runner: '>=2.1.4'
+  build_web_compilers: '>=3.2.1'
+  build_daemon: '>=3.0.0'
+  lints: '>=2.0.0'
 
 # dependency_overrides:
 #  sqflite_common_ffi:
