@@ -4,7 +4,7 @@ publish_to: none
 version: '0.4.0'
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   path: any
