@@ -1,7 +1,7 @@
 name: sqflite_common_ffi
 homepage: https://github.com/tekartik/sqflite/tree/master/sqflite_common_ffi
 description: sqflite ffi based implementation, for desktop and units tests.
-version: 2.3.0+2
+version: 2.3.3
 funding:
   - https://github.com/sponsors/alextekartik
 
@@ -10,17 +10,19 @@ platforms:
   linux:
   macos:
   windows:
+  android:
+  ios:
 
 topics:
   - sql
   - database
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
-  sqlite3: '>=1.11.0 <3.0.0'
-  sqflite_common: '>=2.5.0-1 <4.0.0'
+  sqlite3: '>=2.3.0 <4.0.0'
+  sqflite_common: '>=2.5.4 <4.0.0'
   synchronized: '>=3.0.0 <5.0.0'
   path: '>=1.8.0 <3.0.0'
   meta: '>=1.3.0 <3.0.0'
