@@ -15,7 +15,7 @@ publish_to: none
 version: 1.1.1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   # analyzer: '>=5.2.0'
@@ -27,14 +27,15 @@ dependencies:
   cupertino_icons: '>=0.1.3'
   sqflite: '>=2.0.0-nullsafety.2'
   sqflite_common_ffi:
+  sqflite_common_ffi_async:
   sqflite_example:
     path: ../sqflite/example
   flutter_web_plugins:
     sdk: flutter
   path_provider:
   sqflite_common:
-  dev_test:
-  process_run: '>=0.11.0+2'
+  dev_build:
+  process_run: '>=0.13.3-1'
   path:
   sqflite_common_ffi_web:
 
@@ -61,6 +62,8 @@ dependency_overrides:
     path: ../sqflite_common_ffi
   sqflite_common_ffi_web:
     path: ../packages_web/sqflite_common_ffi_web
+  sqflite_common_ffi_async:
+    path: ../packages/sqflite_common_ffi_async
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
