@@ -32,7 +32,7 @@ void mainExampleApp() {
 /// Sqflite test app
 class SqfliteExampleApp extends StatefulWidget {
   /// test app.
-  const SqfliteExampleApp({Key? key}) : super(key: key);
+  const SqfliteExampleApp({super.key});
   // This widget is the root of your application.
 
   @override
@@ -109,7 +109,7 @@ class _SqfliteExampleAppState extends State<SqfliteExampleApp> {
 /// App home menu page.
 class MyHomePage extends StatefulWidget {
   /// App home menu page.
-  MyHomePage({Key? key, this.title}) : super(key: key) {
+  MyHomePage({super.key, this.title}) {
     _items.add(
         MainItem('Raw tests', 'Raw SQLite operations', route: testRawRoute));
     _items.add(MainItem('Open tests', 'Open onCreate/onUpgrade/onDowngrade',
