@@ -16,7 +16,7 @@ import 'src/common_import.dart';
 /// Manual test page.
 class ManualTestPage extends StatefulWidget {
   /// Test page.
-  const ManualTestPage({Key? key}) : super(key: key);
+  const ManualTestPage({super.key});
 
   @override
   // ignore: library_private_types_in_public_api
@@ -55,10 +55,6 @@ class _ManualTestPageState extends State<ManualTestPage> {
 
   late List<SqfMenuItem> items;
   late List<ItemWidget> itemWidgets;
-
-  Future<bool> pop() async {
-    return true;
-  }
 
   Future<void> _addAndQuery({int? msDelay, bool? noSynchronized}) async {
     // await databaseFactory.debugSetLogLevel(sqfliteLogLevelVerbose);
@@ -219,11 +215,8 @@ class _ManualTestPageState extends State<ManualTestPage> {
       appBar: AppBar(
         title: const Text('Manual tests'),
       ),
-      body: WillPopScope(
-        onWillPop: pop,
-        child: ListView(
-          children: itemWidgets,
-        ),
+      body: ListView(
+        children: itemWidgets,
       ),
     );
   }
@@ -232,7 +225,7 @@ class _ManualTestPageState extends State<ManualTestPage> {
 /// Multiple db test page.
 class MultipleDbTestPage extends StatelessWidget {
   /// Test page.
-  const MultipleDbTestPage({Key? key}) : super(key: key);
+  const MultipleDbTestPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -267,7 +260,7 @@ class MultipleDbTestPage extends StatelessWidget {
 /// Simple db test page.
 class SimpleDbTestPage extends StatefulWidget {
   /// Simple db test page.
-  const SimpleDbTestPage({Key? key, required this.dbName}) : super(key: key);
+  const SimpleDbTestPage({super.key, required this.dbName});
 
   /// db name.
   final String dbName;
@@ -321,8 +314,8 @@ class _SimpleDbTestPageState extends State<SimpleDbTestPage> {
               final db = await _openDatabase();
               final result =
                   firstIntValue(await db.query('test', columns: ['COUNT(*)']));
-              // Temp for nnbd successfull lint
-              if (mounted) {
+              // Temp for nnbd successful lint
+              if (context.mounted) {
                 ScaffoldMessenger.of(context).showSnackBar(SnackBar(
                   content: Text('$result records'),
                   duration: const Duration(milliseconds: 700),
