@@ -1,10 +1,10 @@
 import 'dart:async';
-import 'dart:html' as html;
+import 'package:web/web.dart' as web;
 
 var lines = <String>[];
 var countLineMax = 100;
-var _output = html.querySelector('#output')!;
-var _input = html.querySelector('#input')!;
+var _output = web.document.querySelector('#output')!;
+var _input = web.document.querySelector('#input')!;
 void write(String message) {
   print(message);
   lines.add(message);
@@ -15,7 +15,7 @@ void write(String message) {
 }
 
 void addButton(String text, FutureOr<void> Function() action) {
-  _input.append(html.ButtonElement()
+  _input.append((web.document.createElement('button') as web.HTMLButtonElement)
     ..innerText = text
     ..onClick.listen((event) async {
       await action();
