@@ -1,6 +1,10 @@
 export 'src/database_factory.dart' show createDatabaseFactoryFfiWeb;
+export 'src/debug/debug.dart' show sqliteFfiWebDebugWebWorker;
 export 'src/sqflite_ffi.dart'
-    show databaseFactoryFfiWeb, databaseFactoryFfiWebNoWebWorker;
+    show
+        databaseFactoryFfiWeb,
+        databaseFactoryFfiWebNoWebWorker,
+        databaseFactoryFfiWebBasicWebWorker;
 export 'src/web/load_sqlite.dart'
     show
         SqfliteFfiWebContext,
