@@ -1,2 +1,2 @@
 export 'database_factory_io.dart'
-    if (dart.library.html) 'database_factory_web.dart';
+    if (dart.library.js_interop) 'database_factory_web.dart';
