@@ -1,8 +1,7 @@
-import 'dart:html' as html;
+import 'dart:js_interop';
 
 import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
-import 'package:sqflite_common_ffi_web/src/debug/debug.dart';
 import 'package:sqflite_common_ffi_web/src/sqflite_ffi_impl_web.dart'
     show SqfliteFfiHandlerWeb;
 import 'package:sqflite_common_ffi_web/src/utils.dart';
@@ -12,12 +11,20 @@ import 'package:sqflite_common_ffi_web/src/web/load_sqlite_web.dart'
         SqfliteFfiWebWorkerException,
         defaultSharedWorkerUri;
 import 'package:synchronized/synchronized.dart';
+import 'package:web/web.dart' as web;
 
 import 'import.dart';
 
 /// The Ffi database factory.
 var databaseFactoryFfiWebNoWebWorkerImpl = () {
   return createDatabaseFactoryFfiWeb(noWebWorker: true);
+}();
+
+/// The Ffi database factory with basic worker
+var databaseFactoryFfiWebBasicWorkerImpl = () {
+  return createDatabaseFactoryFfiWeb(
+      // ignore: invalid_use_of_visible_for_testing_member
+      options: SqfliteFfiWebOptions(forceAsBasicWorker: true));
 }();
 
 /// The Ffi database factory.
@@ -59,32 +66,35 @@ DatabaseFactory createDatabaseFactoryFfiWeb(
 
 // Debug database factory web
 bool get _debug => sqliteFfiWebDebugWebWorker;
+var _log = print;
 
 /// Handle method call not to call the web worker.
 Future<dynamic> ffiMethodCallSendToWebWorker(
     FfiMethodCall methodCall, SqfliteFfiWebContext context) async {
   try {
     if (_debug) {
-      print('main_send: $methodCall');
+      // ignore: avoid_print
+      _log('main_send: $methodCall');
     }
     var map = dataToEncodable(methodCall.toDataMap())!;
     var response = await context.sendRawMessage(map);
     if (_debug) {
-      print('main_recv: $response');
+      _log('main_recv: $response');
     }
     return dataFromEncodable(responseToResultOrThrow(response));
   } catch (e, st) {
     if (_debug) {
-      print(e);
-      print(st);
+      _log(e);
+      _log(st);
     }
     if (e is SqfliteFfiWebWorkerException) {
-      html.window.console.error('''
+      web.console.error('''
 An error occurred while initializing the web worker.
 This is likely due to a failure to find the worker javascript file at ${context.options.sharedWorkerUri ?? defaultSharedWorkerUri}
 
 Please check the documentation at https://github.com/tekartik/sqflite/tree/master/packages_web/sqflite_common_ffi_web#setup-binaries to setup the needed binaries.
-''');
+'''
+          .toJS);
     }
     rethrow;
   }
@@ -95,20 +105,20 @@ Future<dynamic> ffiMethodCallHandleNoWebWorker(
     FfiMethodCall methodCall, SqfliteFfiWebContext context) async {
   try {
     if (_debug) {
-      print('handle $methodCall');
+      _log('handle $methodCall');
     }
     dynamic result = await methodCall.rawHandle();
 
     if (_debug) {
-      print('result: $result');
+      _log('result: $result');
     }
 
     // devPrint('result: $result');
     return result;
   } catch (e, st) {
     if (_debug) {
-      print(e);
-      print(st);
+      _log(e);
+      _log(st);
     }
     throw methodCall.wrapAnyExceptionNoIsolate(e);
   }
