@@ -1,1 +1,2 @@
-export 'sqflite_ffi_io.dart' if (dart.library.html) 'sqflite_ffi_web.dart';
+export 'sqflite_ffi_io.dart'
+    if (dart.library.js_interop) 'sqflite_ffi_web.dart';
