@@ -1,17 +1,18 @@
-import 'dart:html';
+import 'dart:async';
+import 'dart:js_interop';
 
 import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
-import 'package:sqflite_common_ffi_web/src/debug/debug.dart';
 import 'package:sqflite_common_ffi_web/src/import.dart';
 import 'package:sqflite_common_ffi_web/src/sqflite_ffi_impl_web.dart'; // ignore: implementation_imports
 import 'package:sqflite_common_ffi_web/src/utils.dart';
+import 'package:web/web.dart' as web;
 
 import 'constants.dart';
 
 bool get _debug => sqliteFfiWebDebugWebWorker; // devWarning(true); // false
 /// Shared worker globals
 var swGlobals = <String, Object?>{};
-
+var _log = print;
 SqfliteFfiWebContext? _swContext;
 SqfliteFfiWebOptions? _swOptions;
 
@@ -20,19 +21,19 @@ var _debugVersion = 2;
 
 var _shw = '/shw$_debugVersion';
 
-Future<void> _handleMessageEvent(Event event) async {
-  var messageEvent = event as MessageEvent;
-  var rawData = messageEvent.data;
-  var port = messageEvent.ports.first;
+void _handleMessageEvent(web.Event event) async {
+  var messageEvent = event as web.MessageEvent;
+  var rawData = messageEvent.data.dartify();
+  var port = messageEvent.ports.toDart.first;
   try {
     if (rawData is String) {
       if (_debug) {
-        print('$_shw receive text message $rawData');
+        _log('$_shw receive text message $rawData');
       }
-      port.postMessage(rawData);
+      port.postMessage(rawData.toJS);
     } else {
       if (_debug) {
-        print('$_shw recv $rawData');
+        _log('$_shw recv $rawData');
       }
       if (rawData is List) {
         var command = rawData[0];
@@ -41,26 +42,26 @@ Future<void> _handleMessageEvent(Event event) async {
           var data = rawData[1] as Map;
           var key = data['key'] as String;
           var value = data['value'] as Object?;
-          print('$_shw $command $key: $value');
+          _log('$_shw $command $key: $value');
           swGlobals[key] = value;
           port.postMessage(null);
         } else if (command == commandVarGet) {
           var data = rawData[1] as Map;
           var key = data['key'] as String;
           var value = swGlobals[key];
-          print('$_shw $command $key: $value');
+          _log('$_shw $command $key: $value');
           port.postMessage({
             'result': {'key': key, 'value': value}
-          });
+          }.jsify());
         } else {
-          print('$_shw $command unknown');
+          _log('$_shw $command unknown');
           port.postMessage(null);
         }
       } else if (rawData is Map) {
         var ffiMethodCall = FfiMethodCall.fromDataMap(rawData);
 
         if (_debug) {
-          print('$_shw method call $ffiMethodCall');
+          _log('$_shw method call $ffiMethodCall');
         }
         if (ffiMethodCall != null) {
           // Fix data
@@ -69,7 +70,7 @@ Future<void> _handleMessageEvent(Event event) async {
           // Init context on first call
           if (_swContext == null) {
             if (_debug) {
-              print('$_shw loading wasm');
+              _log('$_shw loading wasm');
             }
             _swContext = await sqfliteFfiWebLoadSqlite3Wasm(
                 _swOptions ?? SqfliteFfiWebOptions(),
@@ -77,7 +78,11 @@ Future<void> _handleMessageEvent(Event event) async {
             sqfliteFfiHandler = SqfliteFfiHandlerWeb(_swContext!);
           }
           void postResponse(FfiMethodResponse response) {
-            port.postMessage(response.toDataMap());
+            var data = response.toDataMap();
+            if (_debug) {
+              _log('$_shw resp $data ($port)');
+            }
+            port.postMessage(data.jsify());
           }
 
           try {
@@ -88,16 +93,16 @@ Future<void> _handleMessageEvent(Event event) async {
             postResponse(FfiMethodResponse.fromException(e, st));
           }
         } else {
-          print('$_shw $rawData unknown');
+          _log('$_shw $rawData unknown');
           port.postMessage(null);
         }
       } else {
-        print('$_shw $rawData map unknown');
+        _log('$_shw $rawData map unknown');
         port.postMessage(null);
       }
     }
   } catch (e, st) {
-    print('$_shw error caught $e $st');
+    _log('$_shw error caught $e $st');
     port.postMessage(null);
   }
 }
@@ -106,32 +111,63 @@ Future<void> _handleMessageEvent(Event event) async {
 void mainSharedWorker(List<String> args) {
   // sqliteFfiWebDebugWebWorker = devWarning(true);
   if (_debug) {
-    print('$_shw main($_debugVersion)');
+    _log('$_shw main($_debugVersion)');
   }
-
+  var zone = Zone.current;
   try {
-    SharedWorkerGlobalScope.instance.onConnect.listen((event) async {
-      if (_debug) {
-        print('$_shw onConnect()');
-      }
-      var port = (event as MessageEvent).ports.first;
-      port.addEventListener('message', _handleMessageEvent);
-    });
-  } catch (e) {
-    if (_debug) {
-      print('$_shw not in shared worker, trying basic worker');
-    }
-
+    final scope = (globalContext as web.SharedWorkerGlobalScope);
     try {
-      WorkerGlobalScope.instance
-          .addEventListener('message', _handleMessageEvent);
+      var scopeName = scope.name;
+      if (_debug) {
+        _log('$_shw scopeName: $scopeName');
+      }
     } catch (e) {
       if (_debug) {
-        print('$_shw not in shared worker');
+        _log('$_shw scope.name error $e');
       }
     }
+
+    scope.onconnect = (web.Event event) {
+      zone.run(() {
+        if (_debug) {
+          _log('$_shw onConnect()');
+        }
+        var connectEvent = event as web.MessageEvent;
+        var port = connectEvent.ports.toDart[0];
+
+        port.onmessage = (web.MessageEvent event) {
+          zone.run(() {
+            _handleMessageEvent(event);
+          });
+        }.toJS;
+      });
+    }.toJS;
+  } catch (e) {
+    if (_debug) {
+      _log('$_shw not in shared worker, trying basic worker');
+    }
   }
+
+  final scope = (globalContext as web.DedicatedWorkerGlobalScope);
   if (_debug) {
-    print('$_shw main done ($_debugVersion)');
+    _log('$_shw basic worker support');
+  }
+
+  /// Handle basic web workers
+  /// dirty hack
+  try {
+    scope.onmessage = (web.MessageEvent event) {
+      zone.run(() {
+        _handleMessageEvent(event);
+      });
+    }.toJS;
+  } catch (e) {
+    if (_debug) {
+      _log('$_shw not in shared worker error $e');
+    }
+  }
+
+  if (_debug) {
+    _log('$_shw main done ($_debugVersion)');
   }
 }
