@@ -1,2 +1,3 @@
 export 'load_sqlite_common.dart';
-export 'load_sqlite_io.dart' if (dart.library.html) 'load_sqlite_web.dart';
+export 'load_sqlite_io.dart'
+    if (dart.library.js_interop) 'load_sqlite_web.dart';
