@@ -35,8 +35,7 @@ dynamic mockResult(String method) {
 }
 
 class MockDatabase extends SqfliteDatabaseBase {
-  MockDatabase(SqfliteDatabaseOpenHelper openHelper, [String name = 'mock'])
-      : super(openHelper, name);
+  MockDatabase(super.openHelper, [super.name = 'mock']);
 
   int? version;
   List<String> methods = <String>[];
