@@ -1,6 +1,6 @@
 import 'package:sqflite_common/src/database_file_system.dart';
 
-export 'platform_io.dart' if (dart.library.js) 'platform_web.dart';
+export 'platform_io.dart' if (dart.library.js_interop) 'platform_web.dart';
 
 /// IO/web support
 abstract class Platform {
