@@ -4,6 +4,8 @@ import 'package:sqflite_common/src/constant.dart';
 import 'package:sqflite_common/src/database.dart';
 import 'package:sqflite_common/src/database_mixin.dart';
 
+import 'exception.dart';
+
 /// Transaction param, new in transaction v2
 class SqfliteTransactionParam {
   /// null for no transaction
@@ -18,6 +20,9 @@ class SqfliteTransactionParam {
 mixin SqfliteTransactionMixin implements Transaction {
   /// Optional transaction id, depending on the implementation
   int? transactionId;
+
+  /// True if the transaction has already been terminated (rollback or commit)
+  bool? closed;
 }
 
 /// Transaction implementation
@@ -34,7 +39,7 @@ class SqfliteTransaction
   @override
   SqfliteDatabase get db => database;
 
-  /// True if a transaction is successfull
+  /// True if a transaction is successful
   bool? successful;
 
   @override
@@ -47,3 +52,22 @@ class SqfliteTransaction
 /// Special transaction that is run even if a pending transaction is in progress.
 SqfliteTransaction getForcedSqfliteTransaction(SqfliteDatabaseMixin database) =>
     SqfliteTransaction(database)..transactionId = paramTransactionIdValueForce;
+
+/// Internal helpers.
+extension TransactionPrvExt on Transaction {
+  SqfliteTransaction get _txn => this as SqfliteTransaction;
+
+  /// Check if a database is not closed.
+  ///
+  /// Throw an exception if closed.
+  void checkNotClosed() {
+    if (_txn.closed == true) {
+      throw SqfliteDatabaseException('error transaction_closed', null);
+    }
+  }
+
+  /// Mark a transaction as closed.
+  void markClosed() {
+    _txn.closed = true;
+  }
+}
