@@ -1,2 +1,2 @@
 export 'database_factory_ffi_io.dart'
-    if (dart.library.js) 'database_factory_ffi_web.dart';
+    if (dart.library.js_interop) 'database_factory_ffi_web.dart';
