@@ -58,7 +58,7 @@ abstract class SqfliteTestContext {
   bool get isWeb => kSqfliteIsWeb;
 
   /// Set debug mode on
-  @Deprecated('Deb only')
+  @Deprecated('Dev only')
   Future devSetDebugModeOn(bool on);
 
   /// Native (android, ios, macos) only for now
@@ -67,6 +67,9 @@ abstract class SqfliteTestContext {
   /// True if supported.
   /// Not working an Android native, working with ffi impl.
   bool get supportsUri;
+
+  /// Only sqlite_async supports it for now
+  bool get supportsConcurrentRead;
 }
 
 /// sqflite test context mixin.
@@ -77,6 +80,9 @@ mixin SqfliteTestContextMixin implements SqfliteTestContext {
 
   @override
   bool get supportsDeadLock => false;
+
+  @override
+  bool get supportsConcurrentRead => false;
 
   /// FFI implementation is strict
   @override
