@@ -1,4 +1,4 @@
-import 'package:dev_test/package.dart';
+import 'package:dev_build/package.dart';
 import 'package:path/path.dart';
 
 Future main() async {
