@@ -1,1 +1,1 @@
-export 'database_io.dart' if (dart.library.html) 'database_web.dart';
+export 'database_io.dart' if (dart.library.js_interop) 'database_web.dart';
