@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:dev_test/build_support.dart';
+import 'package:dev_build/build_support.dart';
 import 'package:path/path.dart';
 import 'package:process_run/shell.dart';
 
